# Conflicts:
#	libraries/remote_config/src/main/res/xml/remote_config_default.xml

+++ features/content/play/src/main/java/com/tokopedia/play/view/uimodel/recom/PlayLikeInfoUiModel.kt
@@ -21,7 +21,7 @@ data class PlayLikeParamInfoUiModel(
 )
 
 data class PlayLikeStatusInfoUiModel(
-        val totalLike: Int,
+        val totalLike: Long,
         val totalLikeFormatted: String,
         val isLiked: Boolean,
         val source: LikeSource
